requirements.txt
absl-py==2.0.0
alembic==1.12.1
annotated-types==0.5.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astropy==5.3.1
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
autograd==1.6.2
autograd-gamma==0.5.0
autoviz==0.1.730
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
bokeh==2.4.3
cachetools==5.3.2
catboost==1.2.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
colorcet==3.0.1
colorlog==6.7.0
comm==0.1.4
confuse==2.0.1
contourpy==1.1.0
cycler==0.11.0
dask==2023.12.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docx==0.2.4
emoji==2.8.0
executing==1.2.0
fastdtw==0.3.4
fastjsonschema==2.18.0
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.42.0
formulaic==1.0.1
fqdn==1.5.1
fsspec==2023.9.1
future==0.18.3
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.59.2
h5py==3.10.0
holoviews==1.14.9
htmlmin==0.1.12
hvplot==0.7.3
idna==3.4
ImageHash==4.3.1
imbalanced-learn==0.11.0
imblearn==0.0
importlib-metadata==7.0.1
importlib-resources==6.0.0
interface-meta==1.3.0
ipykernel==6.25.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.1.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.25.0
keras==2.14.0
keras-tqdm==2.0.1
kiwisolver==1.4.4
lazypredict==0.2.12
libclang==16.0.6
lifelines==0.28.0
lightgbm==4.1.0
llvmlite==0.41.1
locket==1.0.0
lxml==5.2.2
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
missingno==0.5.2
mistune==3.0.1
ml-dtypes==0.2.0
mlxtend==0.23.1
mpmath==1.2.1
multimethod==1.9.1
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
notebook==7.0.3
notebook_shim==0.2.3
numba==0.58.1
numpy==1.25.2
oauthlib==3.2.2
opt-einsum==3.3.0
optuna==3.4.0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandas-dq==1.28
pandas-profiling==2.7.1
pandocfilters==1.5.0
panel==0.14.4
param==1.13.0
parso==0.8.3
partd==1.4.1
patsy==0.5.3
phik==0.12.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
plotly==5.18.0
polars==0.20.3
ppscore==1.3.0
prettytable==3.9.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.5
psycopg2==2.9.6
pure-eval==0.2.2
pyamg==5.0.1
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyct==0.5.0
pydantic==1.10.12
pydantic_core==2.4.0
pyerfa==2.0.0.3
Pygments==2.15.1
pynndescent==0.5.12
pyparsing==3.0.9
pyphen==0.14.0
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytorch-tabnet==4.1.0
pytz==2023.3
pyviz_comms==3.0.0
PyWavelets==1.4.1
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
rsa==4.9
scikeras==0.12.0
scikit-learn==1.3.0
scikit-multilearn==0.2.0
scipy==1.11.1
seaborn==0.12.2
Send2Trash==1.8.2
shap==0.44.0
six==1.16.0
sklearn-lvq==1.1.1
sklvq==0.1.2
slack_sdk==3.31.0
slicer==0.0.7
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.19
stack-data==0.6.2
statsmodels==0.14.0
sweetviz==2.1.4
sympy==1.11.1
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
textblob==0.17.1
textstat==0.7.3
threadpoolctl==3.2.0
tinycss2==1.2.1
toolz==0.12.0
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
tslearn==0.6.3
typing_extensions==4.12.2
tzdata==2023.3
umap==0.1.1
umap-learn==0.5.6
uri-template==1.3.0
urllib3==2.0.4
visions==0.4.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==3.0.1
widgetsnbextension==4.0.8
wordcloud==1.9.2
wrapt==1.14.1
xgboost==2.0.0
xlrd==2.0.1
zipp==3.17.0
